+++ pkg/client/connector/service_proxy.go
@@ -168,8 +168,12 @@ func (p *mgrProxy) ReviewIntercept(ctx context.Context, arg *managerrpc.ReviewIn
 	return client.ReviewIntercept(ctx, arg, callOptions...)
 }
 
-func (p *mgrProxy) ConnTunnel(_ managerrpc.Manager_ConnTunnelServer) error {
-	return errors.New("ConnTunnel is not implemented by the mgrProxy")
+func (p *mgrProxy) ClientTunnel(_ managerrpc.Manager_ClientTunnelServer) error {
+	return errors.New("ClientTunnel is not implemented by the mgrProxy")
+}
+
+func (p *mgrProxy) AgentTunnel(server managerrpc.Manager_AgentTunnelServer) error {
+	return errors.New("AgentTunnel is not implemented by the mgrProxy")
 }
 
 func (p *mgrProxy) LookupHost(ctx context.Context, arg *managerrpc.LookupHostRequest) (*managerrpc.LookupHostResponse, error) {